Move tests aswell
